ml_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
transformers[torch]==4.35.2
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0+cpu
numpy==1.26.2
scikit-learn==1.3.2
pandas==2.1.3
PyMuPDF==1.23.7
python-dotenv==1.0.0
gunicorn==21.2.0 